Improved site connectivity+more UI upgrades

+++ app/Http/Livewire/FriendRequestList.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Livewire;
 
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Auth;
 use App\Events\RequestAcceptedEvent;
 use App\Events\UserUnfriendedEvent;
@@ -10,36 +11,44 @@ use App\Models\User;
 
 class FriendRequestList extends Component
 {
+    public $user;
 
     public function acceptRequest(int $sender_id, string $notificationId) {
-        $recipient = Auth::user();
-        $sender = User::find($sender_id);
+        if(Gate::allows('private', $this->user->id)){
+            $recipient = Auth::user();
+            $sender = User::find($sender_id);
 
-        $recipient->friends()->attach($sender);
-        $sender->friends()->attach($recipient);
+            $recipient->friends()->attach($sender);
+            $sender->friends()->attach($recipient);
 
-        // Deletes the notification/request and if the sender also had a request from this current user,
-        // that notification is deleted as well.
-        $recipient->notifications()->firstWhere('id', $notificationId)->delete();
-        $sender->notifications()->firstWhere('data->sender_id', $recipient->id)?->delete();
+            // Deletes the notification/request and if the sender also had a request from this current user,
+            // that notification is deleted as well.
+            $recipient->notifications()->firstWhere('id', $notificationId)->delete();
+            $sender->notifications()->firstWhere('data->sender_id', $recipient->id)?->delete();
 
-        RequestAcceptedEvent::dispatch($sender->id, $recipient);
-        $this->emitTo('top-bar', 'friendsUpdated');
+            RequestAcceptedEvent::dispatch($sender->id, $recipient);
+            $this->emitTo('top-bar', 'friendsUpdated');
+        }
     }
 
     public function declineRequest(string $notificationId) {
-        Auth::user()->notifications()->firstWhere('id', $notificationId)->delete();
-        $this->emitTo('top-bar', 'friendsUpdated');
+        if(Gate::allows('private', $this->user->id)){
+            Auth::user()->notifications()->firstWhere('id', $notificationId)->delete();
+            $this->emitTo('top-bar', 'friendsUpdated');
+        }
     }
 
     public function unfriend(User $friend) {
-        $user = Auth::user();
+        if(Gate::allows('private', $this->user->id)){
+            $user = Auth::user();
 
-        $user->friends()->wherePivot('user2_id', $friend->id)->detach();
-        $friend->friends()->wherePivot('user2_id', $user->id)->detach();
+            $user->friends()->wherePivot('user2_id', $friend->id)->detach();
+            $friend->friends()->wherePivot('user2_id', $user->id)->detach();
 
-        UserUnfriendedEvent::dispatch($friend->id, $user);
+            UserUnfriendedEvent::dispatch($friend->id, $user);
+        }
     }
+    
 
     public function render()
     {
